requirements.txt
pydub
requests
python-dotenv
fastapi
uvicorn
python-multipart
chromadb
sentence-transformers
pydub